BiLSTM-RF/requirements.txt
numpy==1.19.5
pandas==1.4.3
scikit-learn==1.1.2
matplotlib==3.5.3
tensorflow-gpu==2.6.0
xgboost==1.5.0
lightgbm==3.3.5